Delete confirmation

+++ src/app/modules/core/pages/dashboard/dashboard.component.ts
@@ -144,7 +144,9 @@ export class DashboardComponent implements OnInit, OnDestroy {
   }
 
   deleteStudent(id: number) {
-    //confirmation
+    if (confirm("Are you sure you want to delete this record?") == false) {
+      return;
+    }
 
     this.studentService.delete(id.toString())
     .pipe(
